requirements.txt
glean-api-client>=0.12.0
